Require `client_id` for SeatGeek API
Fixes #6

SeatGeek will soon require a `client_id` parameter in order to interact with its API. This PR adds support for configuring that to be sent along with HTTP BasicAuth, as well as a query string paramter. Failing to do so will trigger a failure within the Gem alerting the developer to provide the correct credential.